+++ src/main/java/warmup/one/WarmupOneThirdSet.java
@@ -17,4 +17,21 @@ public class WarmupOneThirdSet {
                 || absoluteValueOf200 < 10
                 || absoluteValueOf200 == 10;
     }
+
+    /**
+     * Given 2 int values, return true if one is negative and one is positive. Except if the
+     * parameter "negative" is true, then return true only if both are negative.
+     *
+     * @param a
+     * @param b
+     * @param negative
+     * @return boolean
+     */
+    public boolean posNeg(int a, int b, boolean negative) {
+        if (negative) {
+            return a < 0 && b < 0;
+        }
+
+        return a < 0 && b > 0 || a > 0 && b < 0 ;
+    }
 }